requirements.txt
copernicusmarine
xarray
netCDF4
h5netcdf
h5py
rasterio
numpy
scipy
pandas
dask
mujoco
matplotlib
imageio
msgpack-rpc-python
